chore(separate-concerns): Split Mokha.py into two scripts.
